pyproject.toml
[tool.poetry]
name = "recommendation_toolkit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.20.3"
pandas = "^1.2.4"
torch = "^1.8.1"
python-dotenv = "^0.17.1"
pytorch-lightning = "^1.3.4"
networkx = "^2.5.1"
sklearn = "^0.0"
joblib = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
